+++ gridfs/grid_file.py
@@ -658,7 +658,7 @@ class GridOut(io.IOBase):
     def __iter__(self):
         """Return an iterator over all of this file's data.
 
-        The iterator will return lines (delimited by b'\n') of
+        The iterator will return lines (delimited by ``b'\\n'``) of
         :class:`bytes`. This can be useful when serving files
         using a webserver that handles such an iterator efficiently.